👌 IMPROVE: Improved timetable code structure
