+++ src/components/Timetable/index.jsx
@@ -1,2 +1,4 @@
-export { default as TimetableLectureItem } from './TimetableLectureItem'
 export { default as TimetableContainer } from './TimetableContainer'
+export { default as TimetableCourseItem } from './TimetableCourseItem'
+export { default as TimetableCourseLectureItem } from './TimetableCourseLectureItem'
+export { default as TimetableLayout } from './TimetableLayout'
