+++ src/data/timetableLayoutData.jsx
@@ -1,0 +1,35 @@
+export const cols = [
+  { id: 'col-1', title: 'Monday' },
+  { id: 'col-2', title: 'Tuesday' },
+  { id: 'col-3', title: 'Wednesday' },
+  { id: 'col-4', title: 'Thursday' },
+  { id: 'col-5', title: 'Friday' },
+]
+
+export const rows = [
+  { id: 'time-0830', title: '08:30 am' },
+  { id: 'time-0900', title: '09:00 am' },
+  { id: 'time-0930', title: '09:30 am' },
+  { id: 'time-1000', title: '10:00 am' },
+  { id: 'time-1030', title: '10:30 am' },
+  { id: 'time-1100', title: '11:00 am' },
+  { id: 'time-1130', title: '11:30 am' },
+  { id: 'time-1200', title: '12:00 pm' },
+  { id: 'time-1230', title: '12:30 pm' },
+  // * Lunch break
+  { id: 'time-1400', title: '14:00 pm' },
+  { id: 'time-1430', title: '14:30 pm' },
+  { id: 'time-1500', title: '15:00 pm' },
+  { id: 'time-1530', title: '15:30 pm' },
+  { id: 'time-1600', title: '16:00 pm' },
+  { id: 'time-1630', title: '16:30 pm' },
+  { id: 'time-1700', title: '17:00 pm' },
+  // * Evening break
+  { id: 'time-1730', title: '17:30 pm' },
+  { id: 'time-1800', title: '18:00 pm' },
+  { id: 'time-1830', title: '18:30 pm' },
+  { id: 'time-1900', title: '19:00 pm' },
+  { id: 'time-1930', title: '19:30 pm' },
+  { id: 'time-2000', title: '20:00 pm' },
+  { id: 'time-2030', title: '20:30 pm' },
+]
