+++ src/styles/index.jsx
@@ -1,2 +1,3 @@
 export { default as GlobalStyles } from './Global'
 export { default as themes } from './themes'
+export { default as colorPicker } from './colorPicker'
